(Show Inputs on Overlay) DISABLE_MENU=1 buildfix + Qt settings update
