+++ menu/menu_displaylist.h
@@ -178,7 +178,9 @@ enum menu_displaylist_ctl_state
    DISPLAYLIST_INPUT_MENU_SETTINGS_LIST,
    DISPLAYLIST_LATENCY_SETTINGS_LIST,
    DISPLAYLIST_INPUT_HOTKEY_BINDS_LIST,
+#if defined(HAVE_OVERLAY)
    DISPLAYLIST_ONSCREEN_OVERLAY_SETTINGS_LIST,
+#endif
    DISPLAYLIST_AI_SERVICE_SETTINGS_LIST,
    DISPLAYLIST_ACCESSIBILITY_SETTINGS_LIST,
 #ifdef HAVE_VIDEO_LAYOUT
